create shared util functions
