+++ scripts/commands/create/utils.js
@@ -1,0 +1,95 @@
+const promisify = require('util').promisify
+
+const fs = require('fs')
+const inquirer = require('inquirer')
+const log = require('npmlog')
+const marked = require('meta-marked')
+
+const tutorials = require('../../../src/static/tutorials.json')
+
+const tutorialKeys = Object.keys(tutorials)
+const latestTutorialId = tutorialKeys.sort().reverse()[0]
+const latestTutorial = tutorials[latestTutorialId]
+
+function nextTutorialNumber () {
+  return (parseInt(tutorialKeys.sort()[tutorialKeys.length - 1]) + 1).toString().padStart(4, 0)
+}
+
+// used by lesson, resources
+async function selectTutorial (newItemType) {
+  let tutorial
+  let tutorialId
+  let lessons
+
+  const tutorialResponses1 = await inquirer
+    .prompt([
+      {
+        type: 'confirm',
+        name: 'latestTutorial',
+        message: `Should we add your ${newItemType} to the "${latestTutorial.title}" tutorial?`
+      }
+    ])
+  if (tutorialResponses1.latestTutorial) {
+    tutorial = latestTutorial
+    tutorialId = latestTutorialId
+    lessons = await getTutorialLessons(tutorial, tutorialId)
+  } else {
+    // list existing tutorials + option to create new one
+    let tutorialsList = [{name: 'CREATE NEW TUTORIAL', value: 'new'}]
+    Object.keys(tutorials).sort().forEach(tutorialId => {
+      tutorialsList.push({ name: tutorials[tutorialId].title, value: tutorialId })
+    })
+    const tutorialResponses2 = await inquirer
+      .prompt([
+        {
+          type: 'list',
+          name: 'tutorialId',
+          message: `Which of these existing tutorials should we add your ${newItemType} to?`,
+          choices: tutorialsList.reverse()
+        }
+      ])
+    if (tutorialResponses2.tutorialId !== 'new') {
+      tutorial = tutorials[tutorialResponses2.tutorialId]
+      tutorialId = tutorialResponses2.tutorialId
+      lessons = await getTutorialLessons(tutorial, tutorialId)
+    } else {
+      log.info("I see that you want to create a new tutorial. Someday I'll figure out how to help with that. For now, you'll need to run the command `npm run scripts:create:tutorial` .")
+      // TODO: launch into creating new tutorial
+    }
+  }
+  return { tutorial, tutorialId, lessons }
+}
+
+
+// used by other utils from lessons, resources
+async function getTutorialLessons (tutorial, tutorialId, lessons = [], lessonNumber = 1) {
+  const lessonFilePrefix = `${tutorialId}-${tutorial.url}/${lessonNumber.toString().padStart(2, 0)}`
+
+  let lessonMd
+  let lesson
+  try {
+    lessonMd = await promisify(fs.readFile)(`src/tutorials/${lessonFilePrefix}.md`, 'utf8')
+    lesson = {
+      id: lessonNumber,
+      formattedId: lessonNumber.toString().padStart(2, 0),
+      ...marked(lessonMd).meta
+    }
+  } catch (error) {
+    // lesson not found, we reached the end
+    if (error.code === 'ENOENT') {
+      return lessons
+    }
+
+    // data not well formatted
+    if (error.name === 'YAMLException') {
+      console.error(
+        new Error(`Data improperly formatted in the lesson markdown file "${lessonFilePrefix}.md". Check that the YAML syntax is correct.`)
+      )
+    }
+    throw error
+  }
+  lessons.push(lesson)
+  return getTutorialLessons(tutorial, tutorialId, lessons, lessonNumber + 1)
+}
+
+module.exports = { getTutorialLessons, nextTutorialNumber, selectTutorial }
